skiplist::level_generator

Trait LevelGenerator

pub trait LevelGenerator {
    // Required methods
    fn total(&self) -> usize;
    fn level(&mut self) -> usize;
}

Trait for generating the level of a new node to be inserted into the skiplist.

Required Methods

fn total(&self) -> usize

The total number of levels that are assumed to exist.

fn level(&mut self) -> usize

Generate a random level for a new node in the range [0, total).

This function must not return a level greater or equal to total.