skiplist::ordered_skiplist

Struct OrderedSkipList

pub struct OrderedSkipList<T> { /* private fields */ }

The ordered skiplist provides a way of storing elements such that they are always sorted and at the same time provides efficient way to access, insert and remove nodes. Just like SkipList, it also provides access to indices.

By default, the OrderedSkipList uses the comparison function a.partial_cmp(b).expect("Value cannot be ordered"). This allows the list to handles all types which implement Ord and PartialOrd, though it will panic if value which cannot be ordered is inserted (such as Float::nan()).

The ordered skiplist has an associated sorting function which must be well-behaved. Specifically, given some ordering function f(a, b), it must satisfy the following properties:

Be well defined: f(a, b) should always return the same value

Be anti-symmetric: f(a, b) == Greater iff f(b, a) == Less and f(a, b) == Equal == f(b, a).

By transitive: If f(a, b) == Greater and f(b, c) == Greater then f(a, c) == Greater.

Failure to satisfy these properties can result in unexpected behavior at best, and at worst will cause a segfault, null deref, or some other bad behavior.

Implementations

impl<T> OrderedSkipList<T>
where T: PartialOrd,

pub fn new() -> Self

Create a new skiplist with the default default comparison function of |&a, &b| a.cmp(b).unwrap() and the default number of 16 levels. As a result, any element which cannot be ordered will cause insertion to panic.

The comparison function can always be changed with sort_by, which has essentially no cost if done before inserting any elements.

Panic

The default comparison function will cause a panic if an element is inserted which cannot be ordered (such as Float::nan()).

Examples

use skiplist::OrderedSkipList;

let mut skiplist: OrderedSkipList<i64> = OrderedSkipList::new();

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty skiplist with the optimal number of levels for the intended capacity. Specifically, it uses floor(log2(capacity)) number of levels, ensuring that only a few nodes occupy the highest level.

It uses the default comparison function of |&a, &b| a.cmp(b).unwrap() and can be changed with sort_by.

Panic

The default comparison function will cause a panic if an element is inserted which cannot be ordered (such as Float::nan()).

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::with_capacity(100);
skiplist.extend(0..100);

impl<T> OrderedSkipList<T>
where T: PartialEq,

pub fn index_of(&self, item: &T) -> Option<usize>

Returns the index of the value if it’s contained in the skiplist or None otherwise.

Since this skip list is ordered, this method has an average complexity of O(log n)

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
assert_eq!(skiplist.index_of(&4), 4);

impl<T> OrderedSkipList<T>

pub unsafe fn with_comp<F>(f: F) -> Self
where F: 'static + Fn(&T, &T) -> Ordering,

Create a new skiplist using the provided function in order to determine the ordering of elements within the list. It will be generated with 16 levels.

Safety

The ordered skiplist relies on a well-behaved comparison function. Specifically, given some ordering function f(a, b), it must satisfy the following properties:

Be well defined: f(a, b) should always return the same value

Be anti-symmetric: f(a, b) == Greater if and only if f(b, a) == Less, and f(a, b) == Equal == f(b, a).

By transitive: If f(a, b) == Greater and f(b, c) == Greater then f(a, c) == Greater.

Failure to satisfy these properties can result in unexpected behavior at best, and at worst will cause a segfault, null deref, or some other bad behavior.

Examples

use skiplist::OrderedSkipList;
use std::cmp::Ordering;

// Store even number before odd ones and sort as usual within same parity group.
let mut skiplist = unsafe { OrderedSkipList::with_comp(
    |a: &u64, b: &u64|
    if a%2 == b%2 {
        a.cmp(b)
    } else if a%2 == 0 {
        Ordering::Less
    } else {
        Ordering::Greater
    })};

pub unsafe fn sort_by<F>(&mut self, cmp: F)
where F: 'static + Fn(&T, &T) -> Ordering,

Change the method which determines the ordering of the elements in the skiplist.

Panics

This call will panic if the ordering of the elements will be changed as a result of this new comparison method.

As a result, sort_by is best to call if the skiplist is empty or has just a single element and may panic with 2 or more elements.

Safety

The ordered skiplist relies on a well-behaved comparison function. Specifically, given some ordering function f(a, b), it must satisfy the following properties:

Be well defined: f(a, b) should always return the same value

Be anti-symmetric: f(a, b) == Greater if and only if f(b, a) == Less, and f(a, b) == Equal == f(b, a).

By transitive: If f(a, b) == Greater and f(b, c) == Greater then f(a, c) == Greater.

Failure to satisfy these properties can result in unexpected behavior at best, and at worst will cause a segfault, null deref, or some other bad behavior.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
unsafe { skiplist.sort_by(|a: &i64, b: &i64| b.cmp(a)) } // All good; skiplist empty.
skiplist.insert(0);                                      // Would still be good here.
skiplist.insert(10);
unsafe { skiplist.sort_by(|a: &i64, b: &i64| a.cmp(b)) } // Panics; order would change.

pub fn clear(&mut self)

Clears the skiplist, removing all values.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
skiplist.clear();
assert!(skiplist.is_empty());

pub fn len(&self) -> usize

Returns the number of elements in the skiplist.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
assert_eq!(skiplist.len(), 10);

pub fn is_empty(&self) -> bool

Returns true if the skiplist contains no elements.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
assert!(skiplist.is_empty());

skiplist.insert(1);
assert!(!skiplist.is_empty());

pub fn insert(&mut self, value: T)

Insert the element into the skiplist.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();

skiplist.insert(0);
skiplist.insert(5);
assert_eq!(skiplist.len(), 2);
assert!(!skiplist.is_empty());

pub fn front(&self) -> Option<&T>

Provides a reference to the front element, or None if the skiplist is empty.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
assert!(skiplist.front().is_none());

skiplist.insert(1);
skiplist.insert(2);
assert_eq!(skiplist.front(), Some(&1));

pub fn back(&self) -> Option<&T>

Provides a reference to the back element, or None if the skiplist is empty.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
assert!(skiplist.back().is_none());

skiplist.insert(1);
skiplist.insert(2);
assert_eq!(skiplist.back(), Some(&2));

pub fn get(&self, index: usize) -> Option<&T>

Provides a reference to the element at the given index, or None if the skiplist is empty or the index is out of bounds.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
assert!(skiplist.get(0).is_none());
skiplist.extend(0..10);
assert_eq!(skiplist.get(0), Some(&0));
assert!(skiplist.get(10).is_none());

pub fn pop_front(&mut self) -> Option<T>

Removes the first element and returns it, or None if the sequence is empty.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.insert(1);
skiplist.insert(2);

assert_eq!(skiplist.pop_front(), Some(1));
assert_eq!(skiplist.pop_front(), Some(2));
assert!(skiplist.pop_front().is_none());

pub fn pop_back(&mut self) -> Option<T>

Removes the last element and returns it, or None if the sequence is empty.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.insert(1);
skiplist.insert(2);

assert_eq!(skiplist.pop_back(), Some(2));
assert_eq!(skiplist.pop_back(), Some(1));
assert!(skiplist.pop_back().is_none());

pub fn contains(&self, value: &T) -> bool

Returns true if the value is contained in the skiplist.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
assert!(skiplist.contains(&4));
assert!(!skiplist.contains(&15));

pub fn remove(&mut self, value: &T) -> Option<T>

Removes and returns an element with the same value or None if there are no such values in the skiplist.

If the skiplist contains multiple values with the desired value, the highest level one will be removed. This will results in a deterioration in the skiplist’s performance if the skiplist contains many duplicated values which are very frequently inserted and removed. In such circumstances, the slower remove_first method is preferred.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
assert_eq!(skiplist.remove(&4), Some(4)); // Removes the last one
assert!(skiplist.remove(&4).is_none()); // No more '4' left

pub fn remove_first(&mut self, value: &T) -> Option<T>

Removes and returns an element with the same value or None if there are no such values in the skiplist.

If the skiplist contains multiple values with the desired value, the first one in the skiplist will be returned. If the skiplist contains many duplicated values which are frequently inserted and removed, this method should be preferred over remove.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
for _ in 0..10 {
    skiplist.extend(0..10);
}
assert!(skiplist.remove(&15).is_none());
for _ in 0..9 {
    for i in 0..10 {
        skiplist.remove_first(&i);
    }
}
assert_eq!(skiplist.remove_first(&4), Some(4)); // Removes the last one
assert!(skiplist.remove_first(&4).is_none()); // No more '4' left

pub fn remove_index(&mut self, index: usize) -> T

Removes and returns an element with the given index.

Panics

Panics is the index is out of bounds.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
assert_eq!(skiplist.remove_index(4), 4);
assert_eq!(skiplist.remove_index(4), 5);

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements e such that f(&e) returns false. This method operates in place.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
skiplist.retain(|&x| x%2 == 0);

pub fn dedup(&mut self)

Removes all repeated elements in the skiplist using the skiplist’s comparison function.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..5);
skiplist.insert(3);
skiplist.dedup();

pub fn into_iter(self) -> IntoIter<T>

Get an owning iterator over the entries of the skiplist.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
for i in skiplist.into_iter() {
    println!("Value: {}", i);
}

pub fn iter(&self) -> Iter<'_, T>

Creates an iterator over the entries of the skiplist.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
for i in skiplist.iter() {
    println!("Value: {}", i);
}

pub fn range(&self, min: Bound<&T>, max: Bound<&T>) -> Iter<'_, T>

Constructs a double-ended iterator over a sub-range of elements in the skiplist, starting at min, and ending at max. If min is Unbounded, then it will be treated as “negative infinity”, and if max is Unbounded, then it will be treated as “positive infinity”. Thus range(Unbounded, Unbounded) will yield the whole collection.

Examples

use skiplist::OrderedSkipList;
use std::ops::Bound::{Included, Unbounded};

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
for i in skiplist.range(Included(&3), Included(&7)) {
    println!("Value: {}", i);
}
assert_eq!(Some(&4), skiplist.range(Included(&4), Unbounded).next());

pub fn index_range(&self, range: Range<usize>) -> Iter<'_, T>

Constructs a double-ended iterator over a sub-range of elements in the skiplist, starting at min, and ending at max.

Examples

use skiplist::OrderedSkipList;

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);
for i in skiplist.index_range(3..8) {
    println!("Value: {}", i);
}
assert_eq!(Some(&4), skiplist.index_range(4..5).next());

pub fn lower_bound(&self, min: Bound<&T>) -> Option<&T>

Returns an Option<&T> pointing to the lowest element whose key is above the given bound. If no such element is found then None is returned.

Examples

use skiplist::OrderedSkipList;
use std::ops::Bound::{Included, Excluded, Unbounded};

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);

assert_eq!(skiplist.lower_bound(Unbounded), Some(&0));
assert_eq!(skiplist.lower_bound(Excluded(&0)), Some(&1));
assert_eq!(skiplist.lower_bound(Included(&0)), Some(&0));
assert_eq!(skiplist.lower_bound(Excluded(&10)), None);

pub fn upper_bound(&self, max: Bound<&T>) -> Option<&T>

Returns an Option<&T> pointing to the highest element whose key is above the given bound. If no such element is found then None is returned.

Examples

use skiplist::OrderedSkipList;
use std::ops::Bound::{Included, Excluded, Unbounded};

let mut skiplist = OrderedSkipList::new();
skiplist.extend(0..10);

assert_eq!(skiplist.upper_bound(Unbounded), Some(&9));
assert_eq!(skiplist.upper_bound(Excluded(&9)), Some(&8));
assert_eq!(skiplist.upper_bound(Included(&9)), Some(&9));