skiplist::skiplist

Struct SkipList

pub struct SkipList<T, Gen = Geometric> { /* private fields */ }

SkipList provides a way of storing elements and provides efficient way to access, insert and remove nodes.

Unlike a standard linked list, the skiplist can skip ahead when trying to find a particular index.

Implementations

impl<T> SkipList<T>

pub fn new() -> Self

Create a new skiplist with the default number of 16 levels.

Examples

use skiplist::SkipList;

let mut skiplist: SkipList<i64> = SkipList::new();

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty skiplist with the optimal number of levels for the intended capacity. Specifically, it uses floor(log2(capacity)) number of levels, ensuring that only a few nodes occupy the highest level.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::with_capacity(100);
skiplist.extend(0..100);

impl<T, G> SkipList<T, G>
where G: LevelGenerator,

pub fn new_from_gen(lg: G) -> Self

Constructs a new, empty skip list using a given LevelGenerator

Examples

use skiplist::{SkipList, level_generator::GeometricalLevelGenerator};

let lg = GeometricalLevelGenerator::new(10, 0.5);
let mut skiplist: SkipList<i64> = SkipList::new_from_gen(lg);

pub fn clear(&mut self)

Clears the skiplist, removing all values.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
skiplist.clear();
assert!(skiplist.is_empty());

pub fn len(&self) -> usize

Returns the number of elements in the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
assert_eq!(skiplist.len(), 10);

pub fn is_empty(&self) -> bool

Returns true if the skiplist contains no elements.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
assert!(skiplist.is_empty());

skiplist.push_back(1);
assert!(!skiplist.is_empty());

pub fn insert(&mut self, value: T, index: usize)

Insert the element into the skiplist at the given index, shifting all subsequent nodes down.

Panics

Panics if the insert index is greater than the length of the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();

skiplist.insert(0, 0);
skiplist.insert(5, 1);
assert_eq!(skiplist.len(), 2);
assert!(!skiplist.is_empty());

pub fn push_front(&mut self, value: T)

Insert the element into the front of the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.push_front(1);
skiplist.push_front(2);

pub fn push_back(&mut self, value: T)

Insert the element into the back of the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.push_back(1);
skiplist.push_back(2);

pub fn front(&self) -> Option<&T>

Provides a reference to the front element, or None if the skiplist is empty.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
assert!(skiplist.front().is_none());

skiplist.push_back(1);
skiplist.push_back(2);
assert_eq!(skiplist.front(), Some(&1));

pub fn front_mut(&mut self) -> Option<&mut T>

Provides a mutable reference to the front element, or None if the skiplist is empty.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
assert!(skiplist.front().is_none());

skiplist.push_back(1);
skiplist.push_back(2);
assert_eq!(skiplist.front_mut(), Some(&mut 1));

pub fn back(&self) -> Option<&T>

Provides a reference to the back element, or None if the skiplist is empty.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
assert!(skiplist.back().is_none());

skiplist.push_back(1);
skiplist.push_back(2);
assert_eq!(skiplist.back(), Some(&2));

pub fn back_mut(&mut self) -> Option<&mut T>

Provides a reference to the back element, or None if the skiplist is empty.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
assert!(skiplist.back().is_none());

skiplist.push_back(1);
skiplist.push_back(2);
assert_eq!(skiplist.back_mut(), Some(&mut 2));

pub fn get(&self, index: usize) -> Option<&T>

Provides a reference to the element at the given index, or None if the skiplist is empty or the index is out of bounds.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
assert!(skiplist.get(0).is_none());
skiplist.extend(0..10);
assert_eq!(skiplist.get(0), Some(&0));
assert!(skiplist.get(10).is_none());

pub fn get_mut(&mut self, index: usize) -> Option<&mut T>

Provides a mutable reference to the element at the given index, or None if the skiplist is empty or the index is out of bounds.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
assert!(skiplist.get_mut(0).is_none());
skiplist.extend(0..10);
assert_eq!(skiplist.get_mut(0), Some(&mut 0));
assert!(skiplist.get_mut(10).is_none());

pub fn pop_front(&mut self) -> Option<T>

Removes the first element and returns it, or None if the sequence is empty.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.push_back(1);
skiplist.push_back(2);

assert_eq!(skiplist.pop_front(), Some(1));
assert_eq!(skiplist.pop_front(), Some(2));
assert!(skiplist.pop_front().is_none());

pub fn pop_back(&mut self) -> Option<T>

Removes the last element and returns it, or None if the sequence is empty.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.push_back(1);
skiplist.push_back(2);

assert_eq!(skiplist.pop_back(), Some(2));
assert_eq!(skiplist.pop_back(), Some(1));
assert!(skiplist.pop_back().is_none());

pub fn remove(&mut self, index: usize) -> T

Removes and returns an element with the given index.

Panics

Panics is the index is out of bounds.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
assert_eq!(skiplist.remove(4), 4);
assert_eq!(skiplist.remove(4), 5);

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements e such that f(&e) returns false. This method operates in place.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
skiplist.retain(|&x| x%2 == 0);

pub fn into_iter(self) -> IntoIter<T>

Get an owning iterator over the entries of the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
for i in skiplist.into_iter() {
    println!("Value: {}", i);
}

pub fn iter(&self) -> Iter<'_, T>

Creates an iterator over the entries of the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
for i in skiplist.iter() {
    println!("Value: {}", i);
}

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Creates an mutable iterator over the entries of the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
for i in skiplist.iter_mut() {
    println!("Value: {}", i);
}

pub fn range(&self, min: Bound<usize>, max: Bound<usize>) -> Iter<'_, T>

Constructs a double-ended iterator over a sub-range of elements in the skiplist, starting at min, and ending at max. If min is Unbounded, then it will be treated as “negative infinity”, and if max is Unbounded, then it will be treated as “positive infinity”. Thus range(Unbounded, Unbounded) will yield the whole collection.

Examples

use skiplist::SkipList;
use std::ops::Bound::{Included, Unbounded};

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
for i in skiplist.range(Included(3), Included(7)) {
    println!("Value: {}", i);
}
assert_eq!(Some(&4), skiplist.range(Included(4), Unbounded).next());

pub fn range_mut( &mut self, min: Bound<usize>, max: Bound<usize>, ) -> IterMut<'_, T>

Constructs a mutable double-ended iterator over a sub-range of elements in the skiplist, starting at min, and ending at max. If min is Unbounded, then it will be treated as “negative infinity”, and if max is Unbounded, then it will be treated as “positive infinity”. Thus range(Unbounded, Unbounded) will yield the whole collection.

Examples

use skiplist::SkipList;
use std::ops::Bound::{Included, Unbounded};

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
for i in skiplist.range_mut(Included(3), Included(7)) {
    println!("Value: {}", i);
}
assert_eq!(Some(&mut 4), skiplist.range_mut(Included(4), Unbounded).next());

pub fn lower_bound(&self, min: Bound<usize>) -> Option<&T>

Returns an Option<&T> pointing to the lowest element whose key is above the given bound. If no such element is found then None is returned.

Examples

use skiplist::SkipList;
use std::ops::Bound::{Included, Excluded, Unbounded};

let mut skiplist = SkipList::new();
skiplist.extend(0..10);

assert_eq!(skiplist.lower_bound(Unbounded), Some(&0));
assert_eq!(skiplist.lower_bound(Excluded(0)), Some(&1));
assert_eq!(skiplist.lower_bound(Included(0)), Some(&0));
assert_eq!(skiplist.lower_bound(Excluded(10)), None);

pub fn upper_bound(&self, max: Bound<usize>) -> Option<&T>

Returns an Option<&T> pointing to the highest element whose key is above the given bound. If no such element is found then None is returned.

Examples

use skiplist::SkipList;
use std::ops::Bound::{Included, Excluded, Unbounded};

let mut skiplist = SkipList::new();
skiplist.extend(0..10);

assert_eq!(skiplist.upper_bound(Unbounded), Some(&9));
assert_eq!(skiplist.upper_bound(Excluded(9)), Some(&8));
assert_eq!(skiplist.upper_bound(Included(9)), Some(&9));

impl<T, G> SkipList<T, G>
where T: PartialEq, G: LevelGenerator,

pub fn contains(&self, value: &T) -> bool

Returns true if the value is contained in the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
assert!(skiplist.contains(&4));
assert!(!skiplist.contains(&15));

pub fn index_of(&self, item: &T) -> Option<usize>

Returns the index of the value if it’s contained in the skiplist or None otherwise.

Since this skip list is not ordered, this method has a complexity of O(n)

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.extend(0..10);
assert_eq!(skiplist.index_of(&4), 4);

pub fn dedup(&mut self)

Removes all consecutive repeated elements in the skiplist.

Examples

use skiplist::SkipList;

let mut skiplist = SkipList::new();
skiplist.push_back(0);
skiplist.push_back(0);
assert_eq!(skiplist.len(), 2);
skiplist.dedup();
assert_eq!(skiplist.len(), 1);