Crate skiplist

A skiplist is a list imlementation such that elements are always sorted, and the insertion and deletion of elements are done in O(log(n)).

Conceptually, a skiplist resembles something like:

<head> ----------> [2] --------------------------------------------------> [9] ---------->
<head> ----------> [2] ------------------------------------[7] ----------> [9] ---------->
<head> ----------> [2] ----------> [4] ------------------> [7] ----------> [9] --> [10] ->
<head> --> [1] --> [2] --> [3] --> [4] --> [5] --> [6] --> [7] --> [8] --> [9] --> [10] ->

where we see that a node [x] will have n pointers to other nodes, where n represents which level that node reaches. The idea is that when a node needs to be found (say for insertion), then you start at the highest level and move down. This allows for many of the lower-level nodes to be skipped thus making it faster.

Each skiplist has an associated sorting function. By default, the sorting function is |a, b| a.cmp(b), but any function is permitted, so long as it satisfies the following properties: - It must be consistent: compare(a, b) should always return the same result; - It should be anti-symmetric: If compare(a, b) == Less, then compare(a, b) == Greater. If this is not true, then some unexpected (and probably unsafe) behaviour may happen.

Due to the nature of the skiplist being always sorted, it is not possible to get mutable pointers to the elements of the skiplist as this could leave the skiplist in an inconsistent state. As a result, the IndexMut traits, iter_mut, and get_mut methods and traits are not implemented.

Structs

IntoIter

Iter

SkipList
The skiplist implementation.